requirements.txt
playwright>=1.44
python-dotenv>=1.0.1
pandas>=2.2
openpyxl>=3.1
requests>=2.31
